+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added:
 - Added missing tags
+- Added item config options
 ### Fixed:
 - Fixed stick recipes and drop
 ### Changed: